FUCKING SSL ON ANDROID
Big changes:

made my own ssl factory class to allow all certificates because
otherwise the staging for handybook.com would not work, and wtf?  It
obviously needs to work.

Also implemented said factory in the book a cleaning activity, and then
got a quote and parsed the JSON and then updated a textview.